test(ApplicationTest): 에러 문구 출력 테스트 추가
