+++ src/test/java/lotto/LottoTest.java
@@ -1,7 +1,9 @@
 package lotto;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
 import lotto.model.Lotto;
@@ -12,6 +14,7 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 class LottoTest {
+    private Lotto lotto;
 
     @DisplayName("로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.")
     @Test
@@ -53,6 +56,28 @@ class LottoTest {
                 Arguments.of((Object) new String[]{"12", "15", "-1", "3", "4", "1"})
         );
 
+    }
+
+    boolean isSame(List<Integer> list1, List<Integer> list2) {
+        for (int i = 0; i < list1.size(); i++) {
+
+            if (!list1.get(i).equals(list2.get(i))) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    @DisplayName("로또 번호 오름차순으로 정렬")
+    @Test
+    void sort() {
+        lotto = new Lotto(Arrays.asList(11, 22, 1, 4, 13, 30));
+        List<Integer> list1 = lotto.getNumbers();
+        List<Integer> list2 = Arrays.asList(1, 4, 11, 13, 22, 30);
+        boolean isSameOrder;
+
+        isSameOrder = isSame(list1, list2);
+        assertThat(isSameOrder).isEqualTo(true);
     }
 
 }
